Skeleton API: fix expected ordering in history data retrieval
The front-end and API documentation expect that returned history data
from a /compact-skeleton query has an order to it. The newest entry
(i.e. the one in a live table) should be returned first. This wasn't
enforced using ORDER BY, but relied so far only on Postgres
concatenating results from UNION ALL operators as they are written,
which isn't guaranteed by Postgres. Instead, individual components are
now combined with an explicit order, guaranteeing the newest result to
be first.