Time series: fix small problem with marking active entries
This was spotted thanks to Mihoko Horigome. To compute past versions of
skeletons, reviews, annotations, etc. the TimeSeries namespace provides
functions to create a history index for each data entry. It provides
ordered access to all versions of an element along with its validity
time range represented as a lower and upper bound. All presently active
entries (i.e. entries in the respective live table) are expected to have
an upper bound of null, representing infinity. The history index
creation methods assigns these values, but couldn't rely on the ordering
of the input data before the fix in the last commit. Since this is
available now, it only needs to check the first element as potential
live table entry. With the limited information that was available
before, all entries were checked, and the test to decide whether an
entry is active or not could result in multiple entries being marked as
active. This is fixed now by only testing the first entry. The test
whether an entry is active or not can still be performed without extra
information by testing whether the upper bound is the same or smaller
than the lower bound, because for live table entries, the back-end will
write the edition time as lower bound and creation time as upper bound
as an indicator. This somewhat convoluted semantic is used to save space
in potentially large responses.