Pixi context: mitigate destruction error on unavailable cache entry
A texture cache entry referenced by `outKey` should be available at this
point in the code, but there are multiple reports on a failed
`destroy()` call due to a non-existent texture cache entry. These
reports only appeared recently and because we didn't change any related
code, this might be related to a browser change. To mitigate this
problem until we have a better understanding of the problem, an extra
check is added so that the `destroy()` call is only performed if there
is a valid cache entry.

The `delete` call is not moved inside the conditional, because it might
be that the field itself is still there and has to be removed, even if
the cache value is `undefined`. Calling `delete` on an non-existent
field is fine and doesn't result in an error.