Use plain "git describe" for version information again
Return to the original version representation using "git describe". This
is the most consistent form and it turns it that it can be run also
during Docker builds on DockerHub with the respective build hooks to
ensure all git names are there. This hook is required regardless of
whether we use "git describe" or creating the commit count information
manually. Therefore, the simpler solution is used, i.e. git describe.

The BASE_VERSION is now only used as a fall back if neither Git
environment nor git describe file are present.