Request queue: decouple from UI
Until now the RequestQueue managed the spinner icon in the upper right
corner by itself. To improve the separation of concerns and make it
easier to modify this behavior, the RequestQueue is now an event source
and a client instance listens to its changes and adjusts the spinner
visibility accordingly.