Tracing overlay nodes: fix memory and display problem of confidence text
Instead of creating a texture of the size of the world space text to
still be readable, a small texture is now created and the enclosing
sprite scaled. This makes confidence values appear again (they were gone
for me on a HiDPI screen before). It also fixes the blurryness beyond
zoom level 0 and seems to save some memory as well.

In fact, some test views with a >2000 confidence labels was only
loadable reliably with this patch. The version before would need so much
memory that no new canvas could be created, which was communicated with
an error like this:

  RangeError: Failed to execute 'getImageData' on
  'CanvasRenderingContext2D': Out of memory at ImageData creation

This is fixed now and even much larger test views with 50k edges load
without memory spikes.

Fixes catmaid/CATMAID#1634