Tracing overlay: do tracing mirror requests in parallel
CATMAID's Tracing Overlay supports the use of tracing data mirror
servers which are used for everything but the active skeleton. If a
mirror is used, a second request for the active skeleton nodes in the
current field of view. So far these two request were done serially,
which added unnecessary wait times. This is changed now so that both
requests are sent in parallel. To facilitate this, CATMAID.fetch() now
supports a "parallel" parameter, that, if truthy, causes a new
RequestQueue to be created for each request.

If this is used a lot, it would make sense to make the RequestQueue
implementation more prototypical to safe memory when new instances are
created.