CLI importer: speed up recomputation of materializations
After new data is imported, its materializations (edges, connectors)
need to be recomputed. This is now donw explicitly only for imported
data, which improves performance in typical scenarios (import is small
compared to existing data). If the old behavior of recomputing
everything in the target project should be used in cases where a full
data base is imported, the --update-project-materializations switch can
be used.

Also fix a few logging string formatting bugs.