Tracing overlay nodes: use shared empty array for "no visibility groups"
The common case with visibility groups is that a node isn't part of any.
In these situations, each node has a single empty array associated with
it. To save this heap allocation, a shared empty list is now used. Since
we only run single threaded, this shouldn't be a problem. We also don't
expect writing to this field.