Tracing overlay: fix label font size
Katharina Eichler reported that node labels were rendered in a way too
small font size in the current dev branch. This is the case since the
change in edd03845e40ce, which changed confidence number font scaling,
making the scaled font size needed anymore. This scaled font size
however was used for label rendering as well and not scaling it anymore
made the label font size too small for being rendered in a project space
view. This commit reintroduces a scaled confidence font size, which can
be used for font-size based scaling.