Context help: don't rely on client settings being loaded during URL parsing
No promise for the client settings are queued at this point and
therefore we can't expect the settings to be available. For now, the
initial visibility adjustment is removed to make `dev` usable in
situations of slower connections. It has to be added again based on a
promise that resolves after the client settings are loaded.