CLI importer: don't attempt FK mapping if FK is null
Some foreign keys (FKs) in CATMAID area allowed to be NULL, e.g. parent
node references for root nodes. For those, no attempt in mapping the
reference to a new ID after storage should be attempted, because there
is nothing to map to (the FK is NULL). This fixes a problem when
importing single nodes, which would suddenly get a parent assigned.