Annotations: check class type in annotation-to-id map creation
If there was an annotation named 'annotation' or the name of any other
non-annotation class instance, the query used before was unreliable.
Depending on the order the database returned all results of a passed in
name in the current project, one class instance was last and therefore
the remaining value in the created dictionary. This was observable by
end points that used a name as annotation reference, rather than IDs,
e.g. /{project_id}/annotations/query-targets. Sometimes they would return
results and sometimes not. Luckily, there is no place in the front-end
at the moment, where we use the annotation name directly.

This also makes the annotation search query slightly faster for
databases with many class instances. More safe is it as well, because
the complete escaping is now done by psycopg2.