Landmark widgets: fix transformed skeleton data structure for local skeletons
Adding support for remote skeletons changed the data structure in which
a display transformation stores skeleton references. Before, an object
mapping skeleton IDs to skeleton models was used. This has been replaced
with a simple list of skeleton models (because there is a possibility of
ID conflicts for remote sources). The code for using local skeletons
is now updated as well.