Tracing tool: fix navigator pointerdown listener re-attach
This is only a temporary fix until the event handling of pointer events
in Navigator, TracingTool and TracingOverlay improves. Due to not
overriding onpointerdown, but using the addEventListener()-API,
reattachements of the Navigator listener are not automatically
overridden and we have to remove it manually during binding activation.
This is unfortunately a rather hacky solution.