Neuron similarity: add support for selective similarity matrix creation
The "Configuration" tab in the Neuron Similarity Widget is used to
create new similarity matrices that form the basis for NBLAST
comparisons. So far only a single set of similar skeletons and a single
set for a random skeleton sample have could be specified. It is now
possible to be more selective in terms of which skeletons are seen as
representative of similar pairs.

The UI allows to select between five similarity modes:

1. All skeletons, which is the same as the current behavior,
2. Sub-groups with same suffix, where all skeletons and transformed
skeletons are looked at as similar if they share a common suffix in
their name (taken from neuron name service). A suffix is everything
after the last underscore character in the name an no suffix is treated
as its own group.
3. Pairs with suffixes "_left" and "_right", in which skeletons with the
same name as except for the suffixes _left and _right form a group.
4. Skeletons with the same name and
5. a custom suffix regex.

The last option is in practice probably the most useful one.