CLI exporter: make exclusion priority an option
By default, if both required annotations and exclusion annotations are
provided, the importer will now only exclude a skeleton if it is not
also directly (!) annotated with a sub-annotation of the required
annotations set. This behavior can be disabled an exclusion can be
enforced as soon as a skeleton is annotated with the exclusion
annotation or one of its sub-annotations. To do so, use the
-exclusion-is-final switch.