Neuron similarity: add basic infrastructure for R data caches
This makes it possible to pre-create the R representation of the
simplified neurons from individual projects in Celery workers based on
previously stored files. Wrappers like a management command will be
added separately. This at least saves the server from repeated downloads
and simplification of a common set of neurons or point clouds.