Volume API: use class based view to improve documentation
As noted by @pgunn, the volume API documentation reuses the GET handling
help text also for POST and DELETE request for a particular volume. This
might be confusing or suggest that the API might not be dangerous (which
in fact it is, it is deleting volumes). To fix this, the volume API
entry point has been changed to a class based view with dedicated
documentation for each method.