Pointclouds: sort in database when retrieving pointcloud lists
CATMAID's 3D Viewer shows all available pointclouds in a drop down menu.
To keep the order of the entries constant make it easier readable for
humans, this list is sorted in the front-end by name. With a larger
number of point clouds, this sorting however can become slow. This in
turn makes loading the whole 3D Viewer slow. To fix this, the pointcloud
API accepts now an `order_by` parameter, which can be set to 'id'
(default) or 'name' (used by 3D Viewer) so that a sorted list is already
returned from the back-end. Sorting in the database is much faster,
which also makes loading the 3D Viewer fast again if there are any point
clouds.