Return DRF serialized DateTimeFields in UTC
This was the default behavior until DRF 3.7, but got changed for
consistency with data input:

https://www.django-rest-framework.org/community/release-notes/#370

The recent update of DRF from a version < 3.7 broke some front-end code
that expected UTC. And since we currently don't use these fields for
data input, these fields are explicitly marked as UTC. This fixes the
"(parse error)" table entries in e.g. the Landmark Widget.