3D viewer: fix race condition when adding landmarks
Two calls to showLandmarkGroup() for the same landmark group were able
to both add new meshes with only one of them would be registered for
removal. The reason was that marking a landmark group as added or at
least being worked on happened only after the asynchronous retrieval of
the group information. The check before the asynchronous call would
therefore only work after the first promise resolution.