Landmark widget: make color configurable for each display transformation
The table of active display transformations in the Display tab now
contains a color button for each transformation, allowing to configure
the color for each transformation separately.

The color override for all transformations can still be used, but is now
disabled by default. For now, colors are also only adjustable in the
landmark layer.