Landmark widget: add option to use multiple source/target mapping
The new checkbox "Multiple mappings" in the display tab will display
additional user interface elements that allow the use of additional
source/target landmark group mappings (instead of only a single
source/target mapping). Point matches are computed independently for
each mapping and only merged for the computation of the final
transformation. This is useful mainly to transform skeletons that span
multiple groups.