Settings widget: fix persisting settings for remote CATMAID instances
A deepCopy() is needed, because the settings datastore tests the current
and the new value for equality and only persists to the back-end if the
values are different. If the same array is used for both sides, they
will always be the same and hence no update is send to the back-end.