Neuron similarity: fix storing of query/target object IDs after NBLAST
The actually used query and object IDs are stored after an NBLAST
computation is finished. This is done because the result matrix is
organized by these row and column names plus the original order could be
changed, because of the use of cached information.

The fields used before were wrong.