Treenodes: improve error message when node is part of sampler domain
A node referenced as either a sampler start or end can't be deleted, Our
database constraints prevent it. The error message was however rather
cryptic for users and therefore this commit adds an explicit check for
sampler user of a treenode that is about to be deleted. This has existed
for sampler intervals before, but no checks for domains were performed.
Combining the different checks in one query reduces also the number of
server rountrips for this from two to one.