Admin: fix and improve UI for downsample factors
This UI has always been cumbersome, but recently broke entirely for
custom factors do to Django changes. Fix this and simplify the
interaction between field and widget here.

This is deceptively difficult because:
- MultiValueField and MultiWidget assume list (and *sometimes* tuple)
types are already decompressed. Downsample factors are a list of
Integer3D, and have to be so elsewhere for use as an ArrayField.
- SimpleArrayField also makes assumptions about coupling with its
default widget.
- The SimpleArrayFields used are of Integer3DFormField, which is another
MultiValueField/MultiWidget of list type.

See discussion in CATMAID/catmaid#1827.