Tracing overlay: when stopping event propagation, let Pixi handle it
For some tracing data event handlers we want the first handler that
receives the event an can handles it (as long as it can). All other
handlers shouldn't even receive the event anymore. So far we only
stopped propagation in the DOM tree, but didn't tell Pixi about it. This
caused some pointer events to be triggered on multiple Pixi elements,
e.g. when lines cross or a circle is on aline, causing multiple split
dialogs when clicked with Ctrl + Shift. Using stopPropagation() on the
Pixi event takes care of that, which should now prevent multiple
handlers reacting to the same Pixi event.