Tracing overlay: fix handling click event on multiple lines
Click short-cuts like Ctrl + Shift + click (split line, delete node),
interact with lines. If multiple lines are close to each other, Pixi
calls all hit event handlers. Because we want this behavior (split line)
only to be applied to the fist hit, we now manually keep track if we are
already handling a click event for a line.

A side-effect of this change is that that the split() operation now
returns a promise. This needs to be done for other click handler
actions, too, but for now only the line double click lead to a
problematic behavior.