Neuron similarity: fix runtime computation
Using process_time() is not what we want, because it is only
process-wide timing and we need system-wide timing. Instead we should
use perf_counter(), which includes sleep. The timeit module will do this
for us on Python 3 and will select the most accurate timer also in
future Python versions ona given platform.

See catmaid/CATMAID#1856