Landmark widget: make radio selects show the selected item
Showing the currently selected value makes it much easier to understand
the state of the current display. This is especially important, because
configuring a new landmark transformation consists of multiple selection
from radio select elements.

This adds a new option to the radio select drop down factory method,
which controls what is displayed in the visible select element. So far,
the title of the radio select was shown. This is still the default
('title'). The option 'selected' shows the currently selected item or
"(none)". The option 'title-selected' will show both. Making the same
change to other radio select drop-down menus is as easy as adding this
parameter to the radio select construction.

See catmaid/CATMAID#1856