User front-end model: fix error when doing an update of a particular name
Sanja Jasek discovered a problem with the anonymous user trying to
retrieve information on specific users. The anonymous user isn't allowed
to retrieve user information by default. If this is attempted
regardless, an error was raised when e.g. the Neuron Navigator tries to
display information on locked neurons, more specifically who locked a
neuron. This is fixed now by providing a fallback if no user
information could be retrieved.