Tracing overlay: use default connector type for connectors w/o links
A connector generally has no type, but based on existing relations
CATMAID will give each connector a type to suggest reasonable options
when creating additional links. If there are now links, the connector
type was assumed to be a synaptic connector so far. This however is
confusing if a default connector type is set (e.g. desmosome) and a new
connector without links is created (or an existing one is stripped of
all its links), new links involving it defaulted so far to synaptic
links rather than the default. This is fixed now and the default
connector type set in the Settings Widget is used. Thanks to Sanja Jasek
for pointing this out.