Tracing overlay: prevent mixing connector types when creating new links
The CATMAID front-end assigns types to connectors based on its links to
treenodes (or uses the default connector type), e.g. 'synaptic' or
'desmosome'. When creating new links it was possible so far to add links
to a connector that wouldn't fit to its existing links, even causing the
(non-persisted) connector type to be overridden. This is fixed now by
checking the connector types before adding a new link to a connector
that already has links. If they conflict, a warning is shown and now
link is added. Thanks to Sanja Jasek for bringing this up.