Neuron similarity: don't load query/target object IDs for similarity list
This speeds up loading the full list of NBLAST results in the Neuron
Similarity widget. In addition to the now optional inclusion of
query/target object IDs in the response to a similarity detail request,
the number of valid and invalid query/target object IDs is returned as
well to be able to show at least a little bit of additional information
for each NBLAST query.

This can still be improved with a dedicated SQL query that doesn't load
the query/target object IDs from the database to only count them in case
no object IDs need to be returned (like e.g. from the similarity list in
the Neuron Similarity widget).