Neuron similarity: add Reverse and Top N optios
The reserse options modifies the scoring between a query and a target
object by computing the reverse score, from target to query instead of
the regular forward score. Self-match normalization will refer to the
target if reverse is enabled.

The Top N option allows to reduce the number of computed and stored
scoring values by reducing the target object set for each query object.
It can speed up the computation especially of mean normalization scores
substantially and, as it seems, can maitain good accuracy.

Also add fields for the initial NBLAST request query and target object
IDs. This makes it possible to rerun a query that uses a Top N
configuration.