Fix deletion of reviewed treenodes
If a treenodes is refrenced from another table, the SQL used to delete
the treenode is currently not able to cascade this deletion to those
foreign key references. We currently don't define treenode foreign key
constraints as CASCADE and the DELETE operation itself doesn't attempt
to CASCADE either (not that it would matter if the FKs aren't set up to
CASCADE). Instead, we rely on Django ORM's cascading delete to capture
FK constraints to the deleted object. This however broke down by
manually deleting treenodes through SQL, because the following ORM
delete() call, wouldn't be able to find the referenced treenode anymore
and hence isn't able to do a cascading delete on all referencing
objects.

This is fixed now by not deleting a treeenode anymore explicitly through
SQL and let the Django ORM take care of referential integrity like we
did before 208e31a45d4, which added the explicit SQL delete. This was
done to set the transaction variable "catmaid.user_id" so that our
database triggers can store the last editor ID in the summary table. To
be able to do that, it is however not required to do the delete as well
as part of the SQL. Both, our SQL code and the Django ORM do their
changes in the same transaction. Therefore, setting only the transaction
variable through SQL and letting Django do the delete, does exactly what
we want: communicate the current user ID to the database triggers and
do a cascading deleting for the treenode.