Skeleton API: add extra options to import endpoint
The `POST /{project_id}/skeletons/import` API has now additional
options: The new parameter `skeleton_id` makes it possible to request a
particular skeleton ID during import, just like it is done for neurons
using `neuron_id`. If a skeleton or neuron with this ID exists already,
a new object is created and the existing one is not touched. If an error
should be raised instead, set the `auto_id` parameter to `false`. If
instead the passed in IDs should replace existing data, the `force`
parameter can be set to `true`. Both options apply to both neurons and
skeletons.

To replace an existing skeleton or neuron, the user has to have edit
permission on both the neuron and skeleton class instances as well as
all treenodes of the skeleton.

The default optiosn (`auto_id = True` and `force = False`) result in the
same default behavior as before.

If the passed in IDs are used for skeletons and neurons, both the
concept and location ID sequences are reset to the current maximum ID in
each table. This is needed, because it is also possible to import using
IDs that come after the current sequence state.