Neuron similarity: speed up similarity listing
Instead of computing the length of each array field in Python, do this
in the database. Especially when the array field content is not actually
needed, this speeds up the listing significantly. If the scoring is
required by the front end, the former simple ORM query is still in use.