CLI exporter: improve ergonomics of CLI arguments for selection of exported objects
So far, when providing explicit command line arguments, it was required
to add a "no" prefix to the individual object types if it should not be
exported. E.g. --notreenodes and --nousers. While this reads nice it is
not very intuitive and having two explicit optionsfor a particular type
(e.g.  --treenodes and --notreenodes) clutters the help page.

Therefore the argument definition is changed and now there is only one
argument per type: --treenodes, --connectors, --annotations and --users.
This will explicititly enable the export for those types. The defaults
(when the argument is not provided) stay the same. To explcitly disable
the export of a type `false`, `no`, `f`, `n` and `0` can be provided as
argument, e.g. `-treenodes false` or `--users n`.