Tools: fix ISO date string parsing of fractional seconds
Using `Math.round()` for fractional seconds is wrong. Instead
`Math.floor()` is now used, which a) is semantically correct, because
the respective millisecond fraction begun, but hasn't finished yet, in
terms of the available time resolution and b) is what the browser `Date`
constructor does as well for passed in numeric values.