Admin: allow deletion of complete projects in UI
So far, deleting projects effectively wasn't possible through the admin
UI if there were many objects referencing it (e.g. many treenodes,
ontology items, etc.), because the Django admin view listed all objects
referencing a project before allowing its deletion. Django does this,
because it wants to take care of cascading deletes by itself by default.
This has been changed so that now a simpler confirmation page is shown
that doesn't list all the objects relating to a project. We use the
database's cascading delete for this instead of Django, because it is
much faster for the number of objects that we typically deal with.