Fix Bad Request (400) error on password change
Marta Costa and Rachel Wilson noticed that attempting to change one's
password results in a Bad Request (400) error. It turns out that the
success_url of PassworchChangeView can't be a URL name, but needs to be
a proper URL or a lazily evaluated reverse lookup. The error messages
provided by Django aren't helpful at all in this situation.