Add tools and models to auto-deactivate users after a specified time
Users can now be set to an inactive state after a sepecified amount of
time.  This time is configured for a user group and it apples to users
if they are members of this group. A message can optionally be displayed
as well as contact users. This is configurable as "Group inactivity
periods" in the admin view.

A celery task is added and scheduled to run at midnight every day by
default (given Celery and Celery Beat ar running).

Also add more explicit login error types and allow exceptions to carry a
payload to the front-end in a field called `meta` of the error response.