Skeleton API: fix race condition when merging skeletons
In rare cases a race condition could lead to inconsistent skeleton IDs
in a skeleton that is merged into different skeletons by different
requests at the same time. This is rather unlikely in most manual neuron
tracing workflows, but can be triggered if third party clients perform
merges as well, potentially in an automated fashion.

Thanks to Alex Bates for providing a good example case that triggers
this race condition.