Sampler: fix deletion of sampler w/o optional class instances
The SAMPLER_CREATED_CLASS is by default "sampler created" and is used to
mark treenodes that are created during sampler interval creation. This
class instance is optional and only created if treenodes are created,
which isn't the case when initially only creating a sampler. The
deletion code so far didn't take this optional character into account.

Fixes catmaid/CATMAID#1872