Widget construction: respect all panels when updating content height
So far only the height of the button panel was taken into account when
computing a new height for the widget content. If the widget content was
already exceeding the available height in the window and a panel was
displayed that isn't the button panel (e.g. node filters), the bottom
part of the widget content was hidden and unreachable by scrollbars.

This is fixed now by taking into account all panels for the height
computation. This is done by adding the class 'windowpanel' to each
panel that should be respected.

Fixes catmaid/CATMAID#1871