NBLAST: fix loading of cached data for target skeletons
Cached data should only be attempted to be appended to the general
target object list, if it is actually present. When there was no cached
data available, an error was raised that a `NoneType` couldn't be
appended to the list of target objects.

Also, make the similar query test consistent with the updated target
code in terms of what variable is used for the test.