Graph widget: allow ungrouping one-element groups
Groups with only one element can't be created through the UI, but can
occur if skeletons are removed that are part of a group. Instead of
assuming that all nodes with one element are no groups, the ungrouping
code now checks if the node ID is explicitly marked as a group.