Graph widget: improve handling of merged skeletons
The Graph Widget listens now to skeleton change and skeleton delete
events. For the former, the respective skeleton is reloaded and as
reaction to the latter, the skeleton is removed (including from groups).

This also fixes an issue when loading previously saved JSON files that
contain groups with skeletons that have since been removed.