Tracing overlay: fix queuing of redraw updates
Queuing all node queries as a microtask in the JavaScript event loop
using promises doesn't work well with how the stack viewer is updated on
a section change at the moment. When queued using promises (microtasks),
for some not fully investigated reason, the redraw() call on the tracing
overlay will be executed after the section was changed in the stack
viewer a second time, when having fast consecutive section changes.
Queuing all spatial queries in the regular submitter queue seems to
fix this.

Fixes catmaid/CATMAID#1849