3D viewer: add polyadicity synapse coloring
A new synapse coloring mode, polyadicity, has been added. The number of
partner nodes for each connector is color coded (for synaptic
connectors, this is the number of postsynapses). The colors and ranges
can be configured through the "Polyadicity colors" button in the
"Shading parameters" tab. This is basically a configurable version of an
absolute "N with partner" coloring.