Graph widget: add additional options to GraphML import
The GraphML files generated by Gephi are in a left-handed coordinate
system while CATMAID's graph widget operates in a right-handed
coordinate space. To make it possible to convert between both worlds, a
GraphML import will now show a small dialog that allows to invert the Y
coordinate (default is true).

The dialog also allows to select a preference for "unit_id" fields in
the imported data over "id" fields, because these are exported from
CATMAID as an explicit ID reference if linearized IDs are used (required
by some tools). There is also an option to prefer the "name" field over
the "label" field, to read the base name CATMAID exports.