Landmarks: improve MLS transformation by using reverse point matches
So far, points that were outside of the bounding box of a landmark group
have been transformed using a weighted average of the strict
source->target transformation and the strict target->source
transformation. This has multiple short comings: it requires far more
operations and is therefore slower and it is not correct with respect to
the MLS transformation space: the inverse MLS transformation of a
skeleton using source->target and target->source should be the same, but
it currently isn't. The current mode of operation adds behavior
different from the regular MLS space for all parts of a skeleton that
are outside of the source group bounding box. This is fixed now by using
the inverse point matches as part of a single MLS transformation
together with the regular source->target point matches. Thanks to
@aschampion for comments on this!