Landmark widget: add simple UI for quick pattern-matching based mappings
This adds a "filter" and "pattern" input field for both source and
target. The filter can be used to select candidate groups from the
source and target group list. The pattern is used to remove parts of the
respective group names so that they can be matched up. This is useful to
e.g. match left/right pairs over multiple groups.