NAT: fix use of resample_by parameter
the nat R package works mainly in um space, while CATMAID works with nm.
Therefore, the input to e.g. dotprops() needs to be converted from nm to
um. This was already done before, but using the resample_by parameter,
which also happens to be in nm (matching the skeleton input). The
resample_by parameter should however be used for the resample parameter
if the dotprops() function and needs to be converted into um as well.