NAT: load skeletons directly from DB rather than through HTTP
This improves NBLAST computation time for non-cached objects
significantly and is less load on the server. The current implementation
can be improved in various ways, but provides a good first step.

Especially memory-wise, the current implementation of dotprops cache
file generation is less then ideal. It requires to keep all skeleton
data in memory until the cache file is written. The reason is that
the existing code passes all of the skeleton data in one go to the
dotprops computation function. It would probably be better to go
one-by-one and store results immediately to the cache file.