3D viewer: use CATMAID.FileExporter for file saving
The CATMAID.FileExporter supports stream based saving, which allows for
larger files to be exported. The changes in this commit don't take
advantage of this though, but provides changes necessary to make this
possible in the future.

The function CATMAID.FileExporter.saveAs() function behaves basically
like the global saveAs() function, but will use the streaming API if
explicitly asked or the CATMAID.Client.Settings set this as default.

To not require the caller to assemble all of the data to save in
advance, the function CATMAID.FileExporter.export() with the same
signature as CATMAID.FileExporter.saveAs() returns a FileExporter
instance, to which can be written additional data using its write()
method. Once all data is passed to the exporter, its save() method needs
to be called.