Neuron similarity widget: fix config read-out for similarity matrix creation
If no transformed skeletons are selected as source for the matching set
of skeletons during creation of a new similarity matrix, the respective
selection in the transformed-skeleton drop-down is "none". Therefore,
the test whether a transformation is selected needs to test for "none"
and not only for general truthyness.

Fixes catmaid/CATMAID#1879