User profiles: don't use instance tool visibility defaults as database defaults
If settings.py variables like `settings.PROFILE_SHOW_TRACING_TOOL` are
used as database column defaults in the `UserProfile` model definition,
Django believes a new migration has to be created if the default is
changed in `settings.py`. This shouldn't happen and there is also no
need to: if a new user profile is created when a new user is created,
the tool visibility is already set from the current settings defaults.