Add missing main dependency: sphinx
Sphinx was so far only installed as a documentation dependency as part
of `requirements-doc.txt`. Sphinx needs to be part of this file still,
because we use it to install everything that is needed to build the
documentation on readthedocs.org. However, we also need Sphinx to run
CATMAID, because in the module `custom_rest_swagger_googledoc` we import
it and use it in `get_googledocstring()`, which in turn is used to
render our API documentation. Thanks to @perlman for the pointer!