N5 tile source: update n5-wasm and unhack loading
Update n5-wasm to commit b3fa5007b3ccedf69f00b49cac1735bc2e99d2c1. This
removes the need to request directories and receive HTTP 200 when
testing for the existing of N5 datasets, which caused problems for many
people as this is not a default or advisable configuration for most HTTP
servers.

These changes also allows n5-wasm to be imported directly as a module,
without any editing of the JS or wasm binary output by wasm-pack.