Improve speed of nearest node query
This query is used typically when clicking a skeleton link, because this
makes the view jump to the closest node of this skeleton. Instead of
loading all nodes of the respective skeleton into memory, PostGIS
indexes are now used to query the closest point, by looking at the 100
closest edge centroids to the query location and then using the closest
edge start point as result. This makes queries much faster for larger
skeletons.

Also, in preparation of providing a solution for catmaid/CATMAID#1845,
the API now doesn't require a skeleton ID or neuron ID anymore and can
query the globally closest node in a project.