Upgrade THREE.js from r95 to r104
This is done in preparation of some upcoming potential performance
improvements that build on a recent version of THREE.js. It also fixes a
few issues we experienced around the use of sptrites (e.g. to show
landmarks).

I tested most 3D Viewer functionality, including the Merge/Split dialog
and the Sampler Widget as well as all shading and coloring modes.

A few minor changes had to be made in 3D Viewer to work with this newer
version of THREE.js:

- to update how a render target is set and the scene is cleared,
- to provide a target Vector2 when calling getSize() on a renderer,
- update Detector.js to WebGL.js.