SVG utls: fix style element namespace
So far the namespace was attached to the style element just by setting
the xmlns attribute. This however caused the XML serializer to create
an additional empty xmlns attribute. This wasn't always the case though
and must have changed at some point. The result was that the generated
XML was invalid. By creating the style element explicitly with a
namespace, this is fixed.