3D viewer: add option to export animation of complete history
So far, the number of frames had to be selected manually for a history
export. This isn't really convenient and therefore it is now possible to
export the complete history, regardless of the number of frames. A new
checkbox is shown as part of the animation export dialog if history
animation is enabled: "Complete history". It is enabled by default and
if it is enabled, the number of frames isn't required as input. Instead
the whole history is exported.

Closes catmaid/CATMAID#1806